Fixed several errors in synchronizing server lists.
Added several options to IWebEventEmitterOptions (name, description, verbose).

WebEventEmitter must now be constructed via  WebEventEmitter.Create().
This method is asynchronous and allows waiting
for full initialisation.

Refactored output to a log function. Currently relies on verbose option.
In the future, should be able to defer to a preferred logger.

Fixed a synchronization error in networkSync.

New and revised emit functions (API is *not* final):
- emit(): emits local AND remote but only returns for local (i.e. immediate).
- localEmit(): ONLY emits locally, returns for it.
- remoteEmitOnly: ONLY emits to remote emitters, awaitable.
- remoteEmit: emit to a specific remote server.